Move gatherModuleDocstrings and extractCodeBlocks to new module.
The new module is Cryptol.TypeCheck.Docstrings.
